requirements.txt
Flask       
Jinja2 
Pillow
requests
jsonschema